Consolidation: test_spp.py
